+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 
 - Added `background_rule` to the parameter of `draw_keypoint`.
 - Added exception handling for getting token.
+- Changed the color model in `download_and_read` from BGR to RGB.
 - Fixed to stop outputting binary responses in the debug log.
 
 ## 1.1.0