+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.2.0
 
-Unreleased
+Released 2020-08-20
 
 - Added `get_comparison` and `get_comparisons` to be able to use AnyMotion's new feature, comparison.
 - Added `background_rule` to the parameter of `draw_keypoint`.